切换到rdesktop库
