+++ go.mod
@@ -6,16 +6,13 @@ require (
 	github.com/creack/pty v1.1.17
 	github.com/dustin/go-humanize v1.0.0
 	github.com/go-bindata/go-bindata/v3 v3.1.3
-	github.com/go-vgo/robotgo v0.100.10
 	github.com/gorilla/websocket v1.4.2
 	github.com/kardianos/service v1.2.0
 	github.com/kisielk/errcheck v1.6.0 // indirect
 	github.com/lwch/logging v0.0.0-20211203094902-ecb971ed7cdc
-	github.com/lwch/rdesktop v0.0.0-20211217105819-32436b28a4b2
+	github.com/lwch/rdesktop v0.0.0-20211228105224-a24bee4c086b
 	github.com/lwch/runtime v0.0.0-20190520054850-8c97e19e0c6d
 	github.com/lwch/yaml v0.0.0-20211206085137-772f80e85a26
-	github.com/shirou/gopsutil v3.21.11+incompatible // indirect
-	github.com/yusufpapurcu/wmi v1.2.2 // indirect
 	golang.org/x/lint v0.0.0-20210508222113-6edffad5e616 // indirect
 	golang.org/x/sys v0.0.0-20211216021012-1d35b9e2eb4e
 	golang.org/x/text v0.3.7
